Testing: Reworked testing to allow for cookie authentication. Restructured test suites to facilitate
