+++ src/database.js
@@ -21,10 +21,11 @@ switch (process.env.NODE_ENV.toLowerCase()) {
 
 async function databaseConnect() {
     try {
+        process.env.NODE_ENV.toLowerCase() != "test" ? console.log("connecting to database at url: " + DB_URI) : "";
+        
         // db connection can take time, wait is required
-        console.log("connecting to database at url: " + DB_URI)
         await mongoose.connect(DB_URI);
-        process.env.NODE_ENV.toLowerCase() != "citest" ? console.log("Database connected") : "";
+        process.env.NODE_ENV.toLowerCase() != "test" ? console.log("Database connected") : "";
     } catch (error) {
         console.log(`databaseConnect failed to connect to DB: \n${JSON.stringify(error)}`);
     }
